Fix configuration usage in traits
- Use one trait for accessing the config
- Move HBaseConfig trait variables into Properties
