+++ app/services/DataAccess.scala
@@ -1,5 +1,6 @@
 package services
 
+import play.api.Configuration
 import uk.gov.ons.sbr.models.units.{ EnterpriseUnit, UnitLinks }
 
 import scala.concurrent.Future
